refactor: change `PARAM_PREFIX` => protected and restructure constraints
