+++ truelearn/models/__init__.py
@@ -1,6 +1,6 @@
 """The truelearn.models implements the learner/event model."""
 
-from .base import BaseKnowledgeComponent
+from ._base import BaseKnowledgeComponent
 from ._event import EventModel
 from ._knowledge import KnowledgeComponent, HistoryAwareKnowledgeComponent, Knowledge
 from ._learner import LearnerModel, LearnerMetaWeights
